Restrict pytest-asyncio to <1.0.0 for own CI tests
Looptime is conceptually incompatible with pytest-asyncio>=1.0.0, since the latter adds multi-scoped event loops, and looptime expects the event loop to be strictly function-scoped: it resets the time to 0 for every test, and patches the event loop on every test, so on. This requires a bigger rework of looptime.

Meanwhile, let its own tests run with the older (compatible) versions of pytest-asyncio.

Note that there is NO direct dependency on `pytest-asyncio`, so this restriction cannot be enforced in looptime's own dependencies. There could be different other prodivers of the `event_loop` fixture, in theory, of which pytest-asyncio is only one.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>