Separate event loop patching & configuration & activation for pytest-asyncio>=1.0.0
We are going to have event loops from higher scopes (class, module, package, session), which must be patched at creation, but activated on demand (or not activated with no demand). This, in turn, is needed to have the event loop capable of being activated only for several tests during their life cycles, and not activated for others, and have different configurations for different tests.

Previously —with pytest-asyncio<1.0.0— we only had the function-scoped event loop, where the patching & configuration & activation could be combined, since a single event loop was used for a single test and then discarded, never reused by multiple tests.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>