Drop the unused state management in the pytest plugin
It was a draft for the case when the fixture's scope is stacked — in case the same-named fixture is initialised several times at different scopes (which is wrong by itself).

This approach was never implemented, but the code leftovers sneaked into the codebase: the state of fixtures was persisted, but never used.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>

---

Additionally, there is a bug in pytest since 2019 (6 years), which prevents using this approach: https://github.com/pytest-dev/pytest/issues/5848 — the post-finalizer hook is called multiple times, depending on the number of dependencies of the fixture itself. The fixture finalizer is executed only once, as expected. Either way, this prevents proper accumulation of the stack-like structures based on the number of calls.

In practice, this led to errors like "RuntimeWarning: Fixture 'unused_tcp_port' not found in the cache of scopes" — for all the fixtures of pytest-asyncio specifically — because the stack item was added once, but then removed 2+ times from the stack.

It is easier to drop the assumption that the same-named fixture can under some circumstances be initialized twice at the same time (overlapping). This makes the post-finalizer hook unneeded entirely.

---

A reminder though: this does NOT help put the fixtures under any kind of automated time compaction mode: while the fixture setup can be wrapped properly (because there is a proper hook), the fixture teardown cannot (because there is no hook for the finalizer — only for the post-finalizer, which is too late).

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>